fast-logger: Fix long strings being printed as empty string.
Until now, this program printed only a newline (output should be
lots of "xxxx..."):

    import qualified Data.Text as T
    import           System.Log.FastLogger

    main :: IO ()
    main = do
      set <- newStderrLoggerSet defaultBufSize
      let text = T.pack $ replicate 8161 'x'
      pushLogStrLn set (toLogStr text)

In `(len, next) <- BBE.runBuilder`, `len == 0` doens't mean that there's
nothing to write; it means that no string has has been rendered into the
provided buffer, because the string was already fully present in memory
as a ByteString (this ByteString is contained in `next`).

That ByteString still has to be written out.

Thanks to Andras Slemmer (@exFalso) for help tracking this down.